music and sfx added
Haven't tested it because game kept breaking or not launching. But music and sfx should be placed fine for all scenes, except Main 2.

Not sure if menu music is set up to play properly, couldn't test it. but should keep going through the intro scene untill it's told to stop.

All the audio assets that are going to be used for the demo are in here if needed.